fix: providers modals not opening correctly, move modals to custom folder
