+++ src/lib/stores/oauth-providers.ts
@@ -1,15 +1,15 @@
 import type { Models } from '@appwrite.io/console';
 import type { SvelteComponent } from 'svelte';
 import { writable } from 'svelte/store';
-import Apple from '../../routes/console/project-[project]/auth/appleOAuth.svelte';
-import Auth0 from '../../routes/console/project-[project]/auth/auth0OAuth.svelte';
-import Authentik from '../../routes/console/project-[project]/auth/authentikOAuth.svelte';
-import GitLab from '../../routes/console/project-[project]/auth/gitlabOAuth.svelte';
-import Google from '../../routes/console/project-[project]/auth/googleOAuth.svelte';
-import Main from '../../routes/console/project-[project]/auth/mainOAuth.svelte';
-import Microsoft from '../../routes/console/project-[project]/auth/microsoftOAuth.svelte';
-import Oidc from '../../routes/console/project-[project]/auth/oidcOAuth.svelte';
-import Okta from '../../routes/console/project-[project]/auth/oktaOAuth.svelte';
+import Apple from '../../routes/console/project-[project]/auth/(providers)/appleOAuth.svelte';
+import Auth0 from '../../routes/console/project-[project]/auth/(providers)/auth0OAuth.svelte';
+import Authentik from '../../routes/console/project-[project]/auth/(providers)/authentikOAuth.svelte';
+import GitLab from '../../routes/console/project-[project]/auth/(providers)/gitlabOAuth.svelte';
+import Google from '../../routes/console/project-[project]/auth/(providers)/googleOAuth.svelte';
+import Main from '../../routes/console/project-[project]/auth/(providers)/mainOAuth.svelte';
+import Microsoft from '../../routes/console/project-[project]/auth/(providers)/microsoftOAuth.svelte';
+import Oidc from '../../routes/console/project-[project]/auth/(providers)/oidcOAuth.svelte';
+import Okta from '../../routes/console/project-[project]/auth/(providers)/oktaOAuth.svelte';
 
 export type Provider = Models.Provider & {
     key: string;
