+++ src/routes/console/project-[project]/auth/(providers)/auth0OAuth.svelte
@@ -5,24 +5,24 @@
     import type { Provider } from '$lib/stores/oauth-providers';
     import { sdk } from '$lib/stores/sdk';
     import { onMount } from 'svelte';
-    import { updateOAuth } from './updateOAuth';
+    import { updateOAuth } from '../updateOAuth';
 
     export let provider: Provider;
+    export let show = false;
 
     let appId: string = null;
     let enabled: boolean = null;
     let clientSecret: string = null;
-    let tenantID: string = null;
+    let auth0Domain: string = null;
     let error: string;
 
     onMount(() => {
         appId ??= provider.appId;
         enabled ??= provider.enabled;
-        if (provider.secret) ({ clientSecret, tenantID } = JSON.parse(provider.secret));
+        if (provider.secret) ({ clientSecret, auth0Domain } = JSON.parse(provider.secret));
     });
 
     const projectId = $page.params.project;
-
     const update = async () => {
         const result = await updateOAuth({ projectId, provider, secret, appId, enabled });
 
@@ -34,23 +34,24 @@
     };
 
     $: secret =
-        clientSecret && tenantID ? JSON.stringify({ clientSecret, tenantID }) : provider.secret;
+        clientSecret && auth0Domain
+            ? JSON.stringify({ clientSecret, auth0Domain })
+            : provider.secret;
 </script>
 
-<Modal {error} onSubmit={update} size="big" show on:close>
+<Modal {error} size="big" bind:show onSubmit={update} on:close>
     <svelte:fragment slot="title">{provider.name} OAuth2 Settings</svelte:fragment>
     <FormList>
         <p>
             To use {provider.name} authentication in your application, first fill in this form. For more
             info you can
-            <a class="link" href={provider.docs} target="_blank" rel="noopener noreferrer">
-                visit the docs.
-            </a>
+            <a class="link" href={provider.docs} target="_blank" rel="noopener noreferrer"
+                >visit the docs.</a>
         </p>
         <InputSwitch id="state" bind:value={enabled} label={enabled ? 'Enabled' : 'Disabled'} />
         <InputText
-            id="appID"
-            label="Application (client) ID"
+            id="clientID"
+            label="Client ID"
             autofocus={true}
             placeholder="Enter ID"
             bind:value={appId} />
@@ -58,14 +59,14 @@
             id="secret"
             label="Client Secret"
             placeholder="Enter Client Secret"
-            showPasswordButton
             minlength={0}
+            showPasswordButton
             bind:value={clientSecret} />
         <InputText
-            id="tenant"
-            label="Target Tenant"
-            placeholder="'common','organizations','consumers' or your TenantID"
-            bind:value={tenantID} />
+            id="domain"
+            label="Auth0 Domain"
+            placeholder="Your Auth0 domain"
+            bind:value={auth0Domain} />
         <Alert type="info">
             To complete set up, add this OAuth2 redirect URI to your {provider.name} app configuration.
         </Alert>
@@ -81,7 +82,7 @@
             disabled={(secret === provider.secret &&
                 enabled === provider.enabled &&
                 appId === provider.appId) ||
-                !(appId && clientSecret && tenantID)}
+                !(appId && clientSecret && auth0Domain)}
             submit>Update</Button>
     </svelte:fragment>
 </Modal>
