+++ src/routes/console/project-[project]/auth/(providers)/oidcOAuth.svelte
@@ -5,23 +5,41 @@
     import type { Provider } from '$lib/stores/oauth-providers';
     import { sdk } from '$lib/stores/sdk';
     import { onMount } from 'svelte';
-    import { updateOAuth } from './updateOAuth';
+    import { updateOAuth } from '../updateOAuth';
+
+    const projectId = $page.params.project;
 
     export let provider: Provider;
+    export let show = false;
 
     let appId: string = null;
     let enabled: boolean = null;
     let clientSecret: string = null;
-    let endpoint: string = null;
+    let wellKnownEndpoint: string = null;
+    let authorizationEndpoint: string = null;
+    let tokenEndpoint: string = null;
+    let userinfoEndpoint: string = null;
     let error: string;
+
+    // secret is valid if clientSecret is set and either wellKnownEndpoint or all of the other endpoints are set
+    $: isValidSecret =
+        clientSecret &&
+        (wellKnownEndpoint || (authorizationEndpoint && tokenEndpoint && userinfoEndpoint));
 
     onMount(() => {
         appId ??= provider.appId;
         enabled ??= provider.enabled;
-        if (provider.secret) ({ clientSecret, endpoint } = JSON.parse(provider.secret));
+        if (provider.secret) {
+            ({
+                clientSecret,
+                wellKnownEndpoint,
+                authorizationEndpoint,
+                tokenEndpoint,
+                userinfoEndpoint
+            } = JSON.parse(provider.secret));
+        }
     });
 
-    const projectId = $page.params.project;
     const update = async () => {
         const result = await updateOAuth({ projectId, provider, secret, appId, enabled });
 
@@ -32,11 +50,18 @@
         }
     };
 
-    $: secret =
-        clientSecret && endpoint ? JSON.stringify({ clientSecret, endpoint }) : provider.secret;
+    $: secret = isValidSecret
+        ? JSON.stringify({
+              clientSecret,
+              wellKnownEndpoint,
+              authorizationEndpoint,
+              tokenEndpoint,
+              userinfoEndpoint
+          })
+        : provider.secret;
 </script>
 
-<Modal {error} size="big" show onSubmit={update} on:close>
+<Modal {error} onSubmit={update} size="big" bind:show on:close>
     <svelte:fragment slot="title">{provider.name} OAuth2 Settings</svelte:fragment>
     <FormList>
         <p>
@@ -48,22 +73,38 @@
         <InputSwitch id="state" bind:value={enabled} label={enabled ? 'Enabled' : 'Disabled'} />
         <InputText
             id="appID"
-            label="App ID"
+            label="Client ID"
             autofocus={true}
             placeholder="Enter ID"
             bind:value={appId} />
         <InputPassword
             id="secret"
-            label="App Secret"
-            placeholder="Enter App Secret"
+            label="Client Secret"
+            placeholder="Enter Client Secret"
             minlength={0}
             showPasswordButton
             bind:value={clientSecret} />
         <InputText
-            id="endpoint"
-            label="Endpoint (optional)"
-            placeholder="Your endpoint"
-            bind:value={endpoint} />
+            id="well-known-endpoint"
+            label="Well-Known Endpoint"
+            placeholder="https://example.com/.well-known/openid-configuration"
+            bind:value={wellKnownEndpoint} />
+        <InputText
+            id="authorization-endpoint"
+            label="Authorization Endpoint"
+            placeholder="https://example.com/authorize"
+            bind:value={authorizationEndpoint} />
+        <InputText
+            id="token-endpoint"
+            label="Token Endpoint"
+            placeholder="https://example.com/token"
+            bind:value={tokenEndpoint} />
+        <InputText
+            id="userinfo-endpoint"
+            label="User Info Endpoint"
+            placeholder="https://example.com/userinfo"
+            bind:value={userinfoEndpoint} />
+
         <Alert type="info">
             To complete set up, add this OAuth2 redirect URI to your {provider.name} app configuration.
         </Alert>
@@ -79,7 +120,7 @@
             disabled={(secret === provider.secret &&
                 enabled === provider.enabled &&
                 appId === provider.appId) ||
-                !(appId && clientSecret)}
+                !(appId && isValidSecret)}
             submit>Update</Button>
     </svelte:fragment>
 </Modal>
