+++ src/routes/console/project-[project]/auth/(providers)/oktaOAuth.svelte
@@ -5,37 +5,25 @@
     import type { Provider } from '$lib/stores/oauth-providers';
     import { sdk } from '$lib/stores/sdk';
     import { onMount } from 'svelte';
-    import { updateOAuth } from './updateOAuth';
+    import { updateOAuth } from '../updateOAuth';
 
     const projectId = $page.params.project;
 
     export let provider: Provider;
+    export let show = false;
 
     let appId: string = null;
     let enabled: boolean = null;
     let clientSecret: string = null;
-    let wellKnownEndpoint: string = null;
-    let authorizationEndpoint: string = null;
-    let tokenEndpoint: string = null;
-    let userinfoEndpoint: string = null;
+    let oktaDomain: string = null;
+    let authorizationServerId: string = null;
     let error: string;
-
-    // secret is valid if clientSecret is set and either wellKnownEndpoint or all of the other endpoints are set
-    $: isValidSecret =
-        clientSecret &&
-        (wellKnownEndpoint || (authorizationEndpoint && tokenEndpoint && userinfoEndpoint));
 
     onMount(() => {
         appId ??= provider.appId;
         enabled ??= provider.enabled;
         if (provider.secret) {
-            ({
-                clientSecret,
-                wellKnownEndpoint,
-                authorizationEndpoint,
-                tokenEndpoint,
-                userinfoEndpoint
-            } = JSON.parse(provider.secret));
+            ({ clientSecret, oktaDomain, authorizationServerId } = JSON.parse(provider.secret));
         }
     });
 
@@ -49,18 +37,13 @@
         }
     };
 
-    $: secret = isValidSecret
-        ? JSON.stringify({
-              clientSecret,
-              wellKnownEndpoint,
-              authorizationEndpoint,
-              tokenEndpoint,
-              userinfoEndpoint
-          })
-        : provider.secret;
+    $: secret =
+        clientSecret && oktaDomain && authorizationServerId
+            ? JSON.stringify({ clientSecret, oktaDomain, authorizationServerId })
+            : provider.secret;
 </script>
 
-<Modal {error} onSubmit={update} size="big" show on:close>
+<Modal {error} onSubmit={update} size="big" bind:show on:close>
     <svelte:fragment slot="title">{provider.name} OAuth2 Settings</svelte:fragment>
     <FormList>
         <p>
@@ -84,25 +67,15 @@
             showPasswordButton
             bind:value={clientSecret} />
         <InputText
-            id="well-known-endpoint"
-            label="Well-Known Endpoint"
-            placeholder="https://example.com/.well-known/openid-configuration"
-            bind:value={wellKnownEndpoint} />
+            id="domain"
+            label="Okta Domain"
+            placeholder="dev-1337.okta.com"
+            bind:value={oktaDomain} />
         <InputText
-            id="authorization-endpoint"
-            label="Authorization Endpoint"
-            placeholder="https://example.com/authorize"
-            bind:value={authorizationEndpoint} />
-        <InputText
-            id="token-endpoint"
-            label="Token Endpoint"
-            placeholder="https://example.com/token"
-            bind:value={tokenEndpoint} />
-        <InputText
-            id="userinfo-endpoint"
-            label="User Info Endpoint"
-            placeholder="https://example.com/userinfo"
-            bind:value={userinfoEndpoint} />
+            id="serverID"
+            label="Authorization Server ID"
+            placeholder="default"
+            bind:value={authorizationServerId} />
 
         <Alert type="info">
             To complete set up, add this OAuth2 redirect URI to your {provider.name} app configuration.
@@ -119,7 +92,7 @@
             disabled={(secret === provider.secret &&
                 enabled === provider.enabled &&
                 appId === provider.appId) ||
-                !(appId && isValidSecret)}
+                !(appId && clientSecret && oktaDomain && authorizationServerId)}
             submit>Update</Button>
     </svelte:fragment>
 </Modal>
